Rename component file, adjust fetch call
